LayoutTests/css3/flexbox/column-flex-child-with-overflow-scroll.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
div {
    box-sizing: border-box;
}

.scroll {
    overflow: scroll;
}

.box {
    width: 100px;
    height: 75px;
    padding: 5px;
    border: 2px solid red;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox')">
<div id=log></div>
<div class="flexbox column">
  <div class="box" data-expected-width="100" data-expected-height="75"></div>
  <div class="box scroll" id="box2" data-expected-width="100" data-expected-height="75"></div>
</div>

<div class="flexbox column-reverse">
  <div class="box" data-expected-width="100" data-expected-height="75"></div>
  <div class="box scroll" id="box4" data-expected-width="100" data-expected-height="75"></div>
</div>
</body>
</html>
